Write user session to utmp
Closes: pop-os/cosmic-greeter#239

The utmp file records logged in sessions, remote and local. Most major
display managers write and clean up sessions for logged in users if the
underlying OS and libc support it. The program who and w read the utmp
file, and some software like OpenSSH checks and writes to it as well.

This patch adds utmp support via libc which does all of the hard work.
Implementations set different values for some of the fields, such as
hostname and id. I opted for the obvious and simplest interpretations of
those fields based on what I've seen skimmed the source code of other
display managers.